chore: allow package-specific extra build scripts
If a package has a custom build scripts (`scripts/build.js`), this script
will be run during the main build task.
