+++ scripts/build-utils.js
@@ -5,6 +5,13 @@ const path = require('path');
 
 const PACKAGES_DIR = path.resolve(__dirname, '../packages');
 
+const config = {
+  srcDir: 'src',
+  buildDir: 'lib',
+  jsPattern: '**/*.js',
+  ignorePattern: '**/*.test.js',
+};
+
 function listPackages() {
   return fs
     .readdirSync(PACKAGES_DIR)
@@ -13,6 +20,7 @@ function listPackages() {
 }
 
 module.exports = {
+  config,
   listPackages,
   PACKAGES_DIR,
 };
